Add files via upload
Various math related miniprojects. 

d10estimate.py is a soft bound on the tenth dedekind number, probably the closest anyone ever got, based on a hard upperbound I calculated elsewhere.

Factor.py is a naive factorization implementation.

genprimes7a is a novel non-deterministic prime generator. It generates the quasi-lucas carmichael numbers which are easy to generate (which include the primes as a subset), and then runs certain tests to filter out a subset of the QLCs, leaving only the primes.